requirements.txt
tweepy
pandas
numpy
scikit-learn
tensorflow-gpu
keras
matplotlib
